+++ src/Data/Aeson/ToJSON/Deriving.hs
@@ -6,13 +6,13 @@
 
 -- |
 -- Module:      Data.Time.Format.Typed
--- Copyright:   (c) 20021 Gautier DI FOLCO
+-- Copyright:   (c) 2021-2022 Gautier DI FOLCO
 -- License:     ISC
 -- Maintainer:  Gautier DI FOLCO <gautier.difolco@gmail.com>
 -- Stability:   experimental
 -- Portability: GHC
 --
--- Provide a newtype to be used with DerivingVia to correctly derive <https://hackage.haskell.org/package/aeson-1.5.6.0/docs/Data-Aeson.html#g:7 ToJSON>.
+-- Provide a newtype to be used with DerivingVia to correctly derive <https://hackage.haskell.org/package/aeson-2.1.0.0/docs/Data-Aeson.html#g:7 ToJSON>.
 --
 -- As simple as:
 --
@@ -20,18 +20,27 @@
 -- @
 -- data W = W Int Int
 --   deriving stock (Generic)
---   deriving (ToJSON) via (ModernToJSON W)
+--   deriving (ToJSON) via (GenericallyToJSONToEncoding W)
+-- @
+--
+-- @
+-- data X = X Int Int
+--   deriving stock (Generic)
+--   deriving (ToJSON) via (GenericallyToEncoding X)
 -- @
 module Data.Aeson.ToJSON.Deriving
-  ( ModernToJSON (..),
+  ( GenericallyToJSONToEncoding (..),
+    GenericallyToEncoding (..),
     ToJSON (..),
   )
 where
 
 import Data.Aeson
+import Data.Maybe (fromMaybe)
 import GHC.Generics
 
-newtype ModernToJSON a = ModernToJSON {unModernToJSON :: a}
+-- | Deriving via 'Generic' 'toJSON' and 'toEncoding'.
+newtype GenericallyToJSONToEncoding a = GenericallyToJSONToEncoding {unGenericallyToJSONToEncoding :: a}
   deriving stock (Eq, Show, Generic)
 
 instance
@@ -40,7 +49,31 @@ instance
     GToJSON' Value Zero (Rep a),
     GToJSON' Encoding Zero (Rep a)
   ) =>
-  ToJSON (ModernToJSON a)
+  ToJSON (GenericallyToJSONToEncoding a)
   where
-  toJSON = genericToJSON defaultOptions . unModernToJSON
-  toEncoding = genericToEncoding defaultOptions . unModernToJSON
+  toJSON =
+    genericToJSON defaultOptions
+      . unGenericallyToJSONToEncoding
+  toEncoding =
+    genericToEncoding defaultOptions
+      . unGenericallyToJSONToEncoding
+
+-- | Deriving via 'Generic' 'toEncoding' ('toJSON' is done via 'decode . encode').
+newtype GenericallyToEncoding a = GenericallyToEncoding {unGenericallyToEncoding :: a}
+  deriving stock (Eq, Show, Generic)
+
+instance
+  ( Generic a,
+    ToJSON a,
+    GToJSON' Encoding Zero (Rep a)
+  ) =>
+  ToJSON (GenericallyToEncoding a)
+  where
+  toJSON =
+    fromMaybe (error "GenericallyToEncoding.toJSON: unable to decode encoded Value")
+      . decode
+      . encode
+      . unGenericallyToEncoding
+  toEncoding =
+    genericToEncoding defaultOptions
+      . unGenericallyToEncoding