+++ ChangeLog.md
@@ -2,6 +2,10 @@
 
 ## Unreleased changes
 
+## 0.2.0.0
+ * Drop `ModernToJSON`, use `GenericallyToJSONToEncoding` instead
+ * Introduce `GenericallyToEncoding`
+
 ## 0.1.1.0
  * Set aeson 2 compatibility